Applying Suggestions
